onboarding explanation screens adjusted and integrated to general root navigation graph
