+++ android/src/main/java/io/parity/signer/screens/initial/explanation/OnboardingScreen3.kt
@@ -1,8 +1,9 @@
-package io.parity.signer.screens.initial.onboarding
+package io.parity.signer.screens.initial.explanation
 
 import android.content.res.Configuration
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.MaterialTheme
@@ -12,20 +13,18 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import io.parity.signer.R
 import io.parity.signer.components.base.PageIndicatorLine
-import io.parity.signer.components.base.PrimaryButtonWide
 import io.parity.signer.domain.Callback
 import io.parity.signer.ui.theme.*
 
 
 @Composable
-internal fun OnboardingScreen4(onContinue: Callback) {
+internal fun OnboardingScreen3(onSkip: Callback) {
 	ForceDarkTheme()
 	Column(
 		Modifier
@@ -43,7 +42,7 @@ internal fun OnboardingScreen4(onContinue: Callback) {
 	) {
 		PageIndicatorLine(
 			totalDots = 4,
-			selectedIndex = 4,
+			selectedIndex = 3,
 			modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp),
 		)
 		Row(
@@ -64,9 +63,17 @@ internal fun OnboardingScreen4(onContinue: Callback) {
 				)
 			}
 			Spacer(modifier = Modifier.weight(1f))
+			Text(
+				stringResource(R.string.onboarding_skip),
+				color = MaterialTheme.colors.primary,
+				style = SignerTypeface.LabelS,
+				modifier = Modifier
+					.padding(horizontal = 6.dp)
+					.clickable(onClick = onSkip)
+			)
 		}
 		Text(
-			text = stringResource(R.string.onboarding_header_4),
+			text = stringResource(R.string.onboarding_header_3),
 			color = MaterialTheme.colors.primary,
 			style = SignerTypeface.TitleS,
 			modifier = Modifier.padding(horizontal = 24.dp, vertical = 8.dp)
@@ -77,7 +84,7 @@ internal fun OnboardingScreen4(onContinue: Callback) {
 				.weight(0.2f)
 		)
 		Image(
-			painter = painterResource(id = R.drawable.onboarding_5),
+			painter = painterResource(id = R.drawable.onboarding_4),
 			contentDescription = null,
 			modifier = Modifier
 				.padding(horizontal = 24.dp)
@@ -85,18 +92,6 @@ internal fun OnboardingScreen4(onContinue: Callback) {
 				.weight(0.7f)
 		)
 		Spacer(modifier = Modifier.padding(top = 16.dp))
-	}
-	Box(
-		modifier = Modifier.fillMaxSize(1f),
-		contentAlignment = Alignment.BottomCenter,
-	) {
-		PrimaryButtonWide(
-			label = "continue",
-			activeBackground = Color.White,
-			activeText = MaterialTheme.colors.pink500,
-			modifier = Modifier.padding(24.dp),
-			onClicked = onContinue,
-		)
 	}
 }
 
@@ -106,10 +101,10 @@ internal fun OnboardingScreen4(onContinue: Callback) {
 	showBackground = true, backgroundColor = 0xFF000000,
 )
 @Composable
-private fun PreviewOnboarding4Small() {
+private fun PreviewOnboarding3Small() {
 	SignerNewTheme {
 		Box(modifier = Modifier.size(320.dp, 568.dp)) {
-			OnboardingScreen4({})
+			OnboardingScreen3({})
 		}
 	}
 }
@@ -119,10 +114,10 @@ private fun PreviewOnboarding4Small() {
 	showBackground = true, backgroundColor = 0xFF000000,
 )
 @Composable
-private fun PreviewOnboarding4Big() {
+private fun PreviewOnboarding3Big() {
 	SignerNewTheme {
 		Box(modifier = Modifier.fillMaxSize(1f)) {
-			OnboardingScreen4({})
+			OnboardingScreen3({})
 		}
 	}
 }
