+++ android/src/main/java/io/parity/signer/screens/initial/explanation/OnboardingScreen4.kt
@@ -1,32 +1,31 @@
-package io.parity.signer.screens.initial.onboarding
+package io.parity.signer.screens.initial.explanation
 
 import android.content.res.Configuration
-import androidx.appcompat.app.AppCompatDelegate
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import io.parity.signer.R
 import io.parity.signer.components.base.PageIndicatorLine
+import io.parity.signer.components.base.PrimaryButtonWide
 import io.parity.signer.domain.Callback
 import io.parity.signer.ui.theme.*
 
 
 @Composable
-internal fun OnboardingScreen1(onSkip: Callback) {
+internal fun OnboardingScreen4(onContinue: Callback) {
 	ForceDarkTheme()
 	Column(
 		Modifier
@@ -44,7 +43,7 @@ internal fun OnboardingScreen1(onSkip: Callback) {
 	) {
 		PageIndicatorLine(
 			totalDots = 4,
-			selectedIndex = 1,
+			selectedIndex = 4,
 			modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp),
 		)
 		Row(
@@ -65,17 +64,9 @@ internal fun OnboardingScreen1(onSkip: Callback) {
 				)
 			}
 			Spacer(modifier = Modifier.weight(1f))
-			Text(
-				stringResource(R.string.onboarding_skip),
-				color = MaterialTheme.colors.primary,
-				style = SignerTypeface.LabelS,
-				modifier = Modifier
-					.padding(horizontal = 6.dp)
-					.clickable(onClick = onSkip)
-			)
 		}
 		Text(
-			text = stringResource(R.string.onboarding_header_1),
+			text = stringResource(R.string.onboarding_header_4),
 			color = MaterialTheme.colors.primary,
 			style = SignerTypeface.TitleS,
 			modifier = Modifier.padding(horizontal = 24.dp, vertical = 8.dp)
@@ -86,7 +77,7 @@ internal fun OnboardingScreen1(onSkip: Callback) {
 				.weight(0.2f)
 		)
 		Image(
-			painter = painterResource(id = R.drawable.onboarding_2),
+			painter = painterResource(id = R.drawable.onboarding_5),
 			contentDescription = null,
 			modifier = Modifier
 				.padding(horizontal = 24.dp)
@@ -95,16 +86,17 @@ internal fun OnboardingScreen1(onSkip: Callback) {
 		)
 		Spacer(modifier = Modifier.padding(top = 16.dp))
 	}
-}
-
-@Composable
-fun ForceDarkTheme() {
-	DisposableEffect(key1 = Unit) {
-		val original = AppCompatDelegate.getDefaultNightMode()
-		AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
-		onDispose {
-			AppCompatDelegate.setDefaultNightMode(original)
-		}
+	Box(
+		modifier = Modifier.fillMaxSize(1f),
+		contentAlignment = Alignment.BottomCenter,
+	) {
+		PrimaryButtonWide(
+			label = "continue",
+			activeBackground = Color.White,
+			activeText = MaterialTheme.colors.pink500,
+			modifier = Modifier.padding(24.dp),
+			onClicked = onContinue,
+		)
 	}
 }
 
@@ -114,10 +106,10 @@ fun ForceDarkTheme() {
 	showBackground = true, backgroundColor = 0xFF000000,
 )
 @Composable
-private fun PreviewOnboarding1Small() {
+private fun PreviewOnboarding4Small() {
 	SignerNewTheme {
 		Box(modifier = Modifier.size(320.dp, 568.dp)) {
-			OnboardingScreen1({})
+			OnboardingScreen4({})
 		}
 	}
 }
@@ -127,10 +119,10 @@ private fun PreviewOnboarding1Small() {
 	showBackground = true, backgroundColor = 0xFF000000,
 )
 @Composable
-private fun PreviewOnboarding1Big() {
+private fun PreviewOnboarding4Big() {
 	SignerNewTheme {
 		Box(modifier = Modifier.fillMaxSize(1f)) {
-			OnboardingScreen1({})
+			OnboardingScreen4({})
 		}
 	}
 }
