+++ android/src/main/java/io/parity/signer/ui/RootNavigationGraph.kt
@@ -1,5 +1,6 @@
 package io.parity.signer.ui
 
+import android.util.Log
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -19,6 +20,7 @@ import androidx.navigation.navigation
 import io.parity.signer.domain.Callback
 import io.parity.signer.domain.findActivity
 import io.parity.signer.screens.initial.eachstartchecks.enableEachStartAppFlow
+import io.parity.signer.screens.initial.explanation.OnboardingExplanationScreenFull
 import io.parity.signer.screens.initial.initialUnlockAppScreenFlow
 import io.parity.signer.screens.initial.splash.splashScreen
 import io.parity.signer.screens.initial.termsconsent.TermsConsentScreenFull
@@ -56,9 +58,9 @@ fun RootNavigationGraph(
 			splashScreen(navController)
 			firstTimeOnlyOnboarding(
 				routePath = MainGraphRoutes.firstTimeOnboarding,
-				globalNavController = navController,
+				navController = navController,
 				onNextStepsNavigate = {
-					navController.navigate(MainGraphRoutes.enableAirgapRoute) {
+					navController.navigate(MainGraphRoutes.eachTimeOnboardingRoute) {
 						popUpTo(0)
 					}
 				},
@@ -76,15 +78,16 @@ object MainGraphRoutes {
 	const val splashRoute = "navigation_point_splash"
 	const val firstTimeOnboarding =
 		"navigation_point_once_onboarding"
-	const val enableAirgapRoute = "navigation_point_enable_airgap"
+	const val eachTimeOnboardingRoute = "navigation_point_enable_airgap"
 	const val initialUnlockRoute = "navigation_point_initial_unlock"
 	const val mainScreenRoute = "navigation_main_screen"
 }
 
 
+
 fun NavGraphBuilder.firstTimeOnlyOnboarding(
 	routePath: String,
-	globalNavController: NavHostController,
+	navController: NavHostController,
 	onNextStepsNavigate: Callback,
 ) {
 	navigation(
@@ -92,10 +95,16 @@ fun NavGraphBuilder.firstTimeOnlyOnboarding(
 		startDestination = FirstTimeOnboarding.onboardingExplanationRoute,
 	) {
 		composable(route = FirstTimeOnboarding.onboardingExplanationRoute) {
-
+			OnboardingExplanationScreenFull(navigateNext = {navController.navigate(FirstTimeOnboarding.termsConsentRoute)})
+			LaunchedEffect(Unit) {
+				Log.d(NAVIGATION_TAG, "onboarding explanation opened")
+			}
 		}
 		composable(route = FirstTimeOnboarding.termsConsentRoute) {
 			TermsConsentScreenFull(navigateNextScreen = onNextStepsNavigate)
+			LaunchedEffect(Unit) {
+				Log.d(NAVIGATION_TAG, "terms screen opened")
+			}
 		}
 	}
 }
